Practive composing git commit messages properly
Part of a proper commit message is making sure there is a space
separating the subject line from the body of the message. The subject
line should be limited to 50 characters as a rule of thumb. Subject
line begins with a capital letter and does not end with a period.
Use of the imperative mood in the subject line is standard practice.
Body of commit message is wrapped at 72 characters.